+++ CHANGELOG.md
@@ -8,11 +8,14 @@
 
 ### Improvements
 
+- Added icons for type filter in item views
+
 ### Bugfixes
 
 - Fixed config values not applying correctly
 - Fixed "Search item" context option in recipe panel not selecting item in search view
 - Fixed "Show hidden" option not hiding items
+- Fixed DisplayCategory for items without DisplayCategory (IGUI_ItemCat_Item)
 
 ## 1.5.7 (19.04.2022)